Revert to local libav dependency and fix Rakefile
The homebrew core was fixed to once again allow local dependencies,
so I reverted that part of the formula. Also, `is_snowleopard_64?`
was deprecated in favor of `MacOS.prefer_64_bit?`

The Rakefile failed to compile, and it uninstalled homebrews that
other packages could have dpended on. Because of the way homebrew
upgrades HEAD formulae, and --force is not propagated through the
dependencies, I changed upgrade to an alias for reinstall.