switch to ffmpeg
libav is crashing on me on some high bitrate 10bit h264 files. Strangely
enought this seems to be an osx only problem. If you are interested in
more details check https://bugzilla.libav.org/show_bug.cgi?id=237.

Until this is fixed I'll be using using ffmpeg for the builds.